Disable idempotence checks, update test playbooks.
When working with Non-Debian systems, the idempotence test
fails in the configuration step, because the log2ram installation
script will always create a new log2ram.conf.

Moreover, in order to test the update and remove states, log2ram
must be installed. Update the testing playbooks for these scenarios
to do just that.